drm/omap: do omap_crtc_flush on dev_lastclose
We restore fbdev mode in dev_lastclose(), but we don't wait for the
modeset to be finished. This could cause problems if the modeset
operation is still going on in the background, but a new drm application
is started.

Add omap_crtc_flush() call after restoring the fbdev mode to avoid any
issues.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>