v4l: ti-vps: vip: Cleanup VPDMA lists in close
VPDMA lists needs to be cleaned up to release all the resources
and clear the internal FSM. This is necessay for robustness of the
driver, upon multiple open/close, driver should not keep the VPDMA
hardware in busy state.

Also, maintain a map of all the channels being used by the driver
so that while closing the stream, driver can post abort channel for
all the channels used by the driver since opening.

Mark the channel number as used when before vpdma_add_out_dtd is called
and reinitialize the channel map after cleanup.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>
Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>