v4l: ti-vpe: vpdma: Make list post atomic operation
Writing to the "VPDMA list attribute" register is considered as a list post.
This informs the VPDMA firmware to load the list from the address which
should be taken from the "VPDMA list address" register.

As these two register writes are dependant, it is important that the two writes
happen in atomic manner. This ensures multiple slices (which share same VPDMA)
can post lists asynchronously and all of them point to the correct addresses.

Slightly modified to implementation for the original patch to use
spin_lock instead of mutex as the list post is also called from interrupt context.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>
Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>