v4l: ti-vpe: vip: fix vip_set_clock_enable for multi slice
The vip_set_clock_enable function directly writes to the VIP_CLK_ENABLE
register without reading it first. This register is shared for enabling
clocks for both of the slices. Due to this, enabling one slice would
disable the other. Fix this by updating only the required bits.

Original patch augmanted to also take of the disable path as well.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>
Signed-off-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>