usb: dwc3: dwc3-omap: USB DRVVBUS pinctrl state should be set from wrapper
The USB DRVVBUS pinctrl state should be changed from the wrapper driver
rather than from the PHY driver. While testing suspend with DRD enabled,
it is seen that we miss interrupts since the pin state is switched to
default later by the PHY driver where as the wrapper driver starts using
the pin before that.

To avoid this maintain pin states in wrapper driver.

Signed-off-by: George Cherian <george.cherian@ti.com>
[nsekhar@ti.com: updates to patch description]
Signed-off-by: Sekhar Nori <nsekhar@ti.com>