PM / Voltagedomain: OMAP: Add support for absolute max voltage for a domain
We need to mark the voltage domain absolute maximum rating. This allows
us to setup voltage for a domain from min to max absolute rating.
Unfortunately, it is not possible to solely depend on maximum OPP
allowed for a domain due to the following:
in the case of a ganged voltage rail, where SMPS_x is shared between
VDD_A and VDD_B, setting voltage_tol for VDD_A can prevent VDD_B from
achieving voltage required for the domain. example:
vdd_A requires voltage Z, vdd_B requires voltage Y,
      setting Z-tol to Z + tol will not work when Y is > or < the range.
instead, we look for absolute max for the voltage domains A and B.
setting Z-tol to MAX_A and Y-tol to MAX_B allows voltage ranges which
fall in the intermediate ranges and work for both VDD_A and VDD_B even
if it may be suboptimal for one of them.

If the max range is not provided the mandatory tolerance is used.

Reported-by: Ravikumar Kattekola <rk@ti.com>
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>