ARM: OMAP2+: sleep43xx: Do not use CKE override during DeepSleep
AM437x offers a CKE override register within the control module that
allows the SoC to take control of the CKE line from the EMIF in order to
forcibly hold it low during suspend in order to prevent glitches and
keep the DDR properly in self-refresh.

Previously this was used as described in the suspend and resume path but
more recently it has been shown to cause hangs during suspend, so remove
the toggling of the bits along with all support code as suspend seems to
work better without it.

Tested-by: Keerthy <j-keerthy@ti.com>
Signed-off-by: Dave Gerlach <d-gerlach@ti.com>