ARM: DRA7x: Clock: Fix dpll rate calculation in bypass mode
On DRA7x, clk_get_rate() on a DPLL clk in kernel returns zero
in bypass mode.
This is because the condtional check for bypass mode
in omap2_get_dpll_rate() is missing DRA7xx platform.

Fix this by adding soc_is_dra7xx() to the check. While at it fix
the omap2_init_dpll_parent() as well.

The logic behind the check is

    On POR most of the DPLLs come in bypass mode,
    if a PLL is in bypass at kernel init, the code in
    omap2_get_dpll_rate() will not realize this and will try to calculate
    the clock rate using the multiplier and the divider, resulting in
    errors.

Signed-off-by: Ravikumar Kattekola <rk@ti.com>