ARM: OMAP2+: hwmod: Use spin_lock_irqsave_nested() for locking
Lockdep validator would warn us about possible deadlock situation when
we execute code with hwmods nested. An example for this is on DRA7x platforms
when McASP is configured to use ATL clock as functional clock. In this case
the enable of McASP would trigger an enable for the ATL hwmod in nested
fashion. To help the lockdep to not alert on this we should set unique
subclass for each oh->_lock when we take them. We can cast the oh pointer
as integer for this which will ensure that we have unique but not changing
subclass per oh->_lock use.

Signed-off-by: Peter Ujfalusi <peter.ujfalusi@ti.com>