remoteproc/pruss: fix the device name for PRU1 on AM33xx
The device name used in the match data for the PRU1 core
on AM33xx SoCs has been added incorrectly in the patch,
3b0f8768 ("remoteproc/pruss: use match data instead of
platform data for PRUs"). Fix this typo to reflect the
device name properly, so that the PRU1 can be loaded and
booted successfully. Without this fix, the PRU1 probe
fails with the following trace,
"pru-rproc 4a338000.pru1: missing or incomplete PRU-private data"

Reported-by: Alejandro Hernandez <ajhernandez@ti.com>
Fixes: 3b0f8768 ("remoteproc/pruss: use match data instead of platform data for PRUs")
Signed-off-by: Suman Anna <s-anna@ti.com>