ARM: dts: AM4372: Correct node addresses for PRU-ICSS1 PRU cores
The addresses for the PRU cores in PRU-ICSS on AM437x differs from
that of the corresponding cores on AM33xx family of SoCs. The node
names are incorrectly using the addresses from the AM33xx PRU cores,
even though the reg properties are defined correctly. Fix this typo
in the node addresses.

Signed-off-by: Suman Anna <s-anna@ti.com>