drm/omap: fix page-flip issues
omapdrm's page flip has a few issues:

1) If the fb is removed after page flip has been queued, but before the
page_flip_worker runs, crtc->fb is NULL and the driver will crash.

2) If the page flip has been queued, but page_flip_worker is not yet
ran, and a full mode-set happens, the page_flip_worker ends up
decrementing the ref count of the new fb, set by the mode-set.

To fix these, this patch adds omap_crtc_cancel_page_flip() function,
which doesn't really cancel the page flip (as that would be rather
difficult to do with the current driver), but it does release the queued
fb, and sets the relevant variables to NULL so that the page_flip_worker
knows to skip the flip.

omap_crtc_cancel_page_flip() is called when doing a full mode-set, and
also if the queued fb was removed before page_flip_worker has ran.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>