Input: palmas-pwrbutton - sync with upstream driver
Upstream driver has a bunch of fixes that were done in the course of
multiple review cycles. So sync with it.

There are bunch of fixes that this finally gets in such as:
We missed a shift of the value, resulting in 0 to be always written
to the register bits that control the long key press configuration.

As a result of this, no matter which value we request, we always
configure 0(6 seconds).

This also maintains the change introduced by "input: misc:
palmas-pwrbutton: add missing bustype"

Fixes: c1bc705de029 ("Input: introduce palmas-pwrbutton")
Signed-off-by: Nishanth Menon <nm@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>