(gpio-fan): Add thermal control hooks
Allow gpio-fan to be used as thermal cooling device for platforms that
use GPIO maps to control fans.

As part of this change, we make the shutdown and remove logic the same
as well.

Back port of: https://patchwork.kernel.org/patch/5594731/
reverted in upstream as a result of http://www.spinics.net/lists/linux-next/msg31573.html

Signed-off-by: Nishanth Menon <nm@ti.com>
Acked-by: Eduardo Valentin <edubezval@gmail.com>