i2c: omap: implement bus recovery
If either SCL or SDA are stuck low, we need to
recover the bus using the procedure described
on section 3.1.16 of the I2C specification.

Note that we're trying to implement the procedure
exactly as described by that section. First we
check which line is stuck low, then implement
one or the other procedure. If SDA recovery procedure
fails, we reset our IP in an attempt to make it work.

Signed-off-by: Felipe Balbi <balbi@ti.com>
[nsekhar@ti.com: convert msleep(1) to usleep_range(1000, 2000)]
Signed-off-by: Sekhar Nori <nsekhar@ti.com>