remoteproc: add macros to use with rproc_da_to_va() flags
The rproc_da_to_va() API takes in a flags argument to pass
it onto the individual remoteproc implementations to aid them
in performing proper address translations on internal memory
regions. As this can vary from one implementation to the other,
define some semantics for this flags argument.

A new flags type is encoded into the upper 16 bits along side
the actual value in the lower 16-bits for the flags argument,
to allow different individual implementations to have better
flexibility in interpreting the flags as per their needs.

The current remoteproc ELF loader code is also updated to
use this semantics to pass the different types of flags used
in different invocations of rproc_da_to_va().

Signed-off-by: Suman Anna <s-anna@ti.com>