rpmsg: rpc: fix overwriting of return value in rppc_xlate_buffers
The rppc_xlate_buffers() function is used for translating the host
buffer descriptors to/from the appropriate remote processor pointers
in the message packets being transmitted/received to/from the remote
processor. The function reuses the same loop for unwinding any buffer
translations in case of an error during the Tx path, but happens to
overwrite the initial error value and return the status from the
unwinding operation. This caused the caller to send a buffer with
incorrect/incomplete pointer translations, and creating subsequent
failure paths on the remote processor during the buffer processing.

Fix this by storing away the initial status error and returning it
to the caller. The logic returns only the first error code and not
the values of any subsequent failures.

Reported-by: Sunita Nadampalli <sunitan@ti.com>
Signed-off-by: Sunita Nadampalli <sunitan@ti.com>
Signed-off-by: Suman Anna <s-anna@ti.com>