tty: serial: 8250: omap: fix kernel crash in suspend-to-ram
omap_device infrastructure has a suspend_noirq hook which
runtime suspends all devices late in the suspend cycle.

This leads to a NULL pointer exception in 8250_omap driver
since by that time 8250_dma has already suspended and rxchan
is set to NULL.

Make an explicit check to see if rxchan is NULL in
runtime_{suspend|resume} hooks to fix this.

Signed-off-by: Sekhar Nori <nsekhar@ti.com>