drm/omap: fix stride vs pixel size check
Commit 9ac71bece8679d62734600d9012550b01d2ced46 ("drm/omap: handle
mismatching color format and buffer width") added a check for
incompatible buffer stride and pixel size, but that check is not
correct.

Instead of checking if the buffer byte stride is an integer number of
bytes-per-pixel, it checks if the buffer pixel width is an integer
number of bytes-per-pixel. Which doesn't make any sense, but
accidentally happened to fix the issue we were seeing.

So fix the check.

Signed-off-by: Tomi Valkeinen <tomi.valkeinen@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>