media: ti-vpe: vip: Fix s_fmt behavior
Currently s_fmt is setup to configure the H/W params and the VIP parser
when invoked. This assumes that s_fmt is always called before s_stream.
Although most user application do always call s_fmt before s_stream,
it is not always the case, g-streamer is an example, where if it
finds that reported format g_fmt is acceptable it will happily start
a capture stream.

So here we make sure that the H/W params and the VIP parser are configured
when streaming starts instead.

Signed-off-by: Benoit Parrot <bparrot@ti.com>
Acked-by: Nikhil Devshatwar <nikhil.nd@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>