media: ti-vpe: vip: Stop dma in STREAMOFF
The 'vip_stop_streaming' implementation just disables the IRQs so as
to stop posting descriptors any more. But it does not guarantee
all the current DMAs are stopped immediately. This cleanup happens
only when releasing the stream, i.e. closing the fd.

When application only calls STREAMOFF and still holds the fd open,
The last submitted descriptor would still finish the DMA after
the next frame. This may lead to the memory corruption.

To avoid this, stop all the DMAs currently pending in the 'stop_streaming'
ioctl itself.

Signed-off-by: Nikhil Devshatwar <nikhil.nd@ti.com>
Acked-by: Benoit Parrot <bparrot@ti.com>
Signed-off-by: Jyri Sarha <jsarha@ti.com>